update all tsconfigs (yes, by hand) it's pretty darn similar except…
the package's tsconfig.json, which i added because it seemed to help the IDE understand things (or else i created a micro supsersitition that led me to that belief) and that file has too many asterisks for some of the typescript problems i worked through. but it does seem more better. like mo' bettah. but i didn't put it in this file because i don't understand what's happening and am therefore, not convinced that it should be included in the recipe of "good tsconfig stuff" yet